Part1, Part2 (not debugged)

+++ libft/ft_strcpy.c
@@ -6,7 +6,7 @@
 /*   By: fujitaharuki <fujitaharuki@student.42.f    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/07 12:00:14 by fujitaharuk       #+#    #+#             */
-/*   Updated: 2024/04/07 23:45:12 by fujitaharuk      ###   ########.fr       */
+/*   Updated: 2024/04/10 06:32:54 by fujitaharuk      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strcpy(char *dest, const char *src)
 {
-	int	i;
+	size_t	i;
 
 	i = -1;
 	while (src[++i])
