+++ libft/ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: fujitaharuki <fujitaharuki@student.42.f    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/07 11:53:41 by fujitaharuk       #+#    #+#             */
-/*   Updated: 2024/04/07 23:32:20 by fujitaharuk      ###   ########.fr       */
+/*   Updated: 2024/04/10 06:30:47 by fujitaharuk      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 char	*ft_strdup(const char *str)
 {
 	char	*dest;
-	int		i;
+	size_t	i;
 
 	dest = (char *)malloc(sizeof(char) * (ft_strlen(str) + 1));
 	i = -1;
