+++ libft/ft_strncpy.c
@@ -6,7 +6,7 @@
 /*   By: fujitaharuki <fujitaharuki@student.42.f    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/07 12:05:17 by fujitaharuk       #+#    #+#             */
-/*   Updated: 2024/04/07 23:44:51 by fujitaharuk      ###   ########.fr       */
+/*   Updated: 2024/04/10 06:31:54 by fujitaharuk      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strncpy(char *dst, const char *str, size_t len)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] && i < len)
